requirements.txt
pyqt5
pillow
matplotlib
